Refactor to library and add integration test
