+++ Cargo.toml
@@ -6,3 +6,6 @@ edition = "2021"
 [dependencies]
 rusqlite = { version = "0.31", features = ["bundled"] }
 rayon = "1.8"
+
+[dev-dependencies]
+tempfile = "3"
